headerのロジックとviewを分離
